chore: jump code improved for maintainability
The concept of a trigger for an ability was added. The first trigger
added was on the jump ability. The jump code follows an internal state
machine using an enum. When not active the jump ability is in a
'grounded' state. In the grounded state, the abilities apply method is a
no-op. When triggered, the state is moved to 'jumping', which allows the
internal state machine to progress until the floor is detected which
returns the ability back to 'grounded'.

In this iteration, jumping is only permitted when the ability is in the
'grounded' state.